United FireArms and Grenades mission logics
